requirements.txt
streamlit
numpy
tensorflow
Pillow
scikit-learn
opencv-python
streamlit-webrtc

